requirements.txt
asgiref==3.3.1
Django==3.1.4
django-resized==0.3.11
Pillow==8.0.1
pkg-resources==0.0.0
pytz==2020.5
sqlparse==0.4.1
